add code permission app home project rest framework

+++ home/views.py
@@ -7,6 +7,7 @@ from rest_framework.decorators import APIView, api_view
 from .serializers import PersonSerializer, QuestionSerializer, AnswerSerializer
 from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
 from rest_framework import status
+from permissions import IsOwnerOrReadOnly
 
 
 class HomeView(APIView):
@@ -18,13 +19,17 @@ class HomeView(APIView):
         return Response(ser_data.data)
 
 
-class QuestionView(APIView):
-    permission_classes = [AllowAny]
+class QuestionListView(APIView):
+    permission_classes = [IsAuthenticated]
 
     def get(self, request):
         question = Question.objects.all()
         ser_data = QuestionSerializer(instance=question, many=True).data
         return Response(ser_data, status=status.HTTP_200_OK)
+
+
+class QuestionCreateView(APIView):
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         ser_data = QuestionSerializer(data=request.data)
@@ -33,17 +38,27 @@ class QuestionView(APIView):
             return Response(ser_data.data, status=status.HTTP_201_CREATED)
         
         return Response(ser_data.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    
+class QuestionUpdateView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
 
     def put(self, request, pk):
         question = Question.objects.get(pk=pk)
+        self.check_object_permissions(request, question)
         ser_data = QuestionSerializer(instance=question, data=request.data, partial=True)
         if ser_data.is_valid():
             ser_data.save()
             return Response(ser_data.data, status=status.HTTP_201_CREATED)
         
         return Response(ser_data.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    
+class QuestionDeleteView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
 
     def delete(self, request, pk):
         question = Question.objects.get(pk=pk)
+        self.check_object_permissions(request, question)
         question.delete()
         return Response({"message":{"question deleted successfully"}})
